Unnecessary boxing

+++ src/main/java/gregtech/api/items/GT_EnergyArmor_Item.java
@@ -162,7 +162,7 @@ public class GT_EnergyArmor_Item extends ItemArmor implements ISpecialArmor {
                 }
             }
 
-            jumpChargeMap.put(aPlayer, Float.valueOf(jumpCharge));
+            jumpChargeMap.put(aPlayer, jumpCharge);
         }
 
         if ((mSpecials & 256) != 0) {
